apps/api/requirements.txt
aio-pika==9.4.1
alembic==1.13.1
boto3==1.34.23
fastapi==0.110.0
httpx==0.26.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.1.18
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
redis==5.0.1
sse-starlette==2.0.0
sqlalchemy==2.0.25
uvicorn[standard]==0.27.0
